requirements.txt
accelerate==0.34.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
brotli==1.0.9
certifi==2024.2.2
charset-normalizer==2.0.4
click==8.1.7
comm==0.2.2
datasets==2.19.2
debugpy==1.6.7
decorator==5.1.1
decord==0.6.0
deepspeed==0.15.4
dill==0.3.8
dnspython==2.7.0
docker-pycreds==0.4.0
docstring-parser==0.16
einops==0.7.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
gitdb==4.0.11
gitpython==3.1.43
gmpy2==2.1.2
hjson==3.1.0
huggingface-hub==0.27.0
idna==3.4
intel-openmp==2024.1.0
ipykernel==6.29.3
ipython==8.22.2
ipywidgets==8.1.5
jedi==0.19.1
jinja2==3.1.3
jupyter-client==7.3.4
jupyter-core==5.7.2
jupyterlab-widgets==3.0.13

markdown-it-py==3.0.0
markupsafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl==2024.1.0
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-ml-py
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
packaging==24.0
pandas==2.2.2
parso==0.8.4
peft==0.10.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.42
protobuf==4.25.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.6.4
pydantic-core==2.16.3
pygments==2.17.2
pynvml==11.5.0
pysocks==1.7.1
python-dateutil==2.9.0
pytz==2024.1
pyyaml==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.32.3
rich==13.7.1
safetensors==0.4.5
sentencepiece==0.2.0
sentry-sdk==1.45.0
setproctitle==1.3.3
shtab==1.7.1
six==1.16.0
smmap==5.0.1
stack-data==0.6.2
sympy==1.13.1
tbb==2021.12.0
tensorboardx==2.6.2.2
tokenizers==0.20.1
torch==2.5.1
torchao==0.6.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.66.4
traitlets==5.14.2
transformers==4.46.3
triton==3.1.0
trl==0.8.6
typing-extensions==4.9.0
tyro==0.8.4
tzdata==2024.1
ujson==5.10.0
urllib3==2.1.0
wandb==0.16.6
wcwidth==0.2.13
widgetsnbextension==4.0.13
xxhash==3.4.1
yarl==1.9.4
gradio==5.9.1